+++ indexed-profunctors/indexed-profunctors.cabal
@@ -1,0 +1,30 @@
+name:          indexed-profunctors
+version:       0.1
+license:       BSD3
+license-file:  LICENSE
+build-type:    Simple
+cabal-version: 1.24
+maintainer:    optics@well-typed.com
+author:        Adam Gundry, Andres Löh, Andrzej Rybczak, Oleg Grenrus
+tested-with:   GHC ==8.0.2 || ==8.2.2 || ==8.4.4 || ==8.6.5 || ==8.8.1, GHCJS ==8.4
+synopsis:      Utilities for indexed profunctors
+category:      Data, Optics, Lenses, Profunctors
+description:
+  This package contains basic definitions related to indexed profunctors.  These
+  are primarily intended as internal utilities to support the @optics@ and
+  @generic-lens@ package families.
+
+bug-reports:   https://github.com/well-typed/optics/issues
+source-repository head
+  type:     git
+  location: https://github.com/well-typed/optics.git
+  subdir:   indexed-profunctors
+
+library
+  default-language: Haskell2010
+  hs-source-dirs:   src
+  ghc-options:      -Wall
+
+  build-depends: base                   >= 4.9        && <5
+
+  exposed-modules:    Data.Profunctor.Indexed
