+++ optics-core/src/Optics/AffineTraversal.hs
@@ -60,9 +60,9 @@ module Optics.AffineTraversal
   )
   where
 
-import Optics.Internal.Concrete
+import Data.Profunctor.Indexed
+
 import Optics.Internal.Optic
-import Optics.Internal.Profunctor
 
 -- | Type synonym for a type-modifying affine traversal.
 type AffineTraversal s t a b = Optic An_AffineTraversal NoIx s t a b
