+++ optics-core/src/Optics/IxGetter.hs
@@ -23,10 +23,11 @@ module Optics.IxGetter
   , A_Getter
   ) where
 
+import Data.Profunctor.Indexed
+
 import Optics.Internal.Bi
 import Optics.Internal.Indexed
 import Optics.Internal.Optic
-import Optics.Internal.Profunctor
 
 -- | Type synonym for an indexed getter.
 type IxGetter i s a = Optic' A_Getter (WithIx i) s a
