+++ optics-core/src/Optics/IxLens.hs
@@ -44,9 +44,10 @@ module Optics.IxLens
 
 import Data.Void
 
+import Data.Profunctor.Indexed
+
 import Optics.Internal.Indexed
 import Optics.Internal.Optic
-import Optics.Internal.Profunctor
 
 -- | Type synonym for a type-modifying indexed lens.
 type IxLens i s t a b = Optic A_Lens (WithIx i) s t a b
