+++ optics-core/src/Optics/IxSetter.hs
@@ -57,10 +57,12 @@ module Optics.IxSetter
   , FunctorWithIndex(..)
   ) where
 
+import Data.Profunctor.Indexed
+
 import Optics.Internal.Indexed
 import Optics.Internal.IxSetter
 import Optics.Internal.Optic
-import Optics.Internal.Profunctor
+import Optics.Internal.Utils
 
 -- | Type synonym for a type-modifying indexed setter.
 type IxSetter i s t a b = Optic A_Setter (WithIx i) s t a b
