+++ optics-extra/optics-extra.cabal
@@ -30,6 +30,7 @@ library
                , bytestring             >= 0.10.8    && <0.11
                , containers             >= 0.5.7.1   && <0.7
                , hashable               >= 1.1.1     && <1.4
+               , indexed-profunctors    >= 0.1       && <0.2
                , mtl                    >= 2.2.2     && <2.3
                , optics-core            >= 0.1       && <0.1.1
                , text                   >= 1.2       && <1.3
