+++ optics/optics.cabal
@@ -125,6 +125,7 @@ test-suite optics-tests
 
   build-depends: base
                , containers
+               , indexed-profunctors    >= 0.1        && <0.2
                , inspection-testing     >= 0.4.1.1    && <0.5
                , mtl
                , optics
